Split decoders to files from parquet_common.
