+++ be/src/vec/exec/format/parquet/parquet_common.cpp
@@ -23,8 +23,6 @@
 
 namespace doris::vectorized {
 
-const cctz::time_zone DecodeParams::utc0 = cctz::utc_time_zone();
-
 const uint32_t ParquetInt96::JULIAN_EPOCH_OFFSET_DAYS = 2440588;
 const uint64_t ParquetInt96::MICROS_IN_DAY = 86400000000;
 const uint64_t ParquetInt96::NANOS_PER_MICROSECOND = 1000;
@@ -187,551 +185,5 @@ size_t ColumnSelectVector::get_next_run(DataReadType* data_read_type) {
         }
         return run_length;
     }
-}
-
-Status Decoder::get_decoder(tparquet::Type::type type, tparquet::Encoding::type encoding,
-                            std::unique_ptr<Decoder>& decoder) {
-    switch (encoding) {
-    case tparquet::Encoding::PLAIN:
-        switch (type) {
-        case tparquet::Type::BOOLEAN:
-            decoder.reset(new BoolPlainDecoder());
-            break;
-        case tparquet::Type::BYTE_ARRAY:
-            decoder.reset(new ByteArrayDecoder());
-            break;
-        case tparquet::Type::INT32:
-        case tparquet::Type::INT64:
-        case tparquet::Type::INT96:
-        case tparquet::Type::FLOAT:
-        case tparquet::Type::DOUBLE:
-        case tparquet::Type::FIXED_LEN_BYTE_ARRAY:
-            decoder.reset(new FixLengthDecoder(type));
-            break;
-        default:
-            return Status::InternalError("Unsupported type {}(encoding={}) in parquet decoder",
-                                         tparquet::to_string(type), tparquet::to_string(encoding));
-        }
-        break;
-    case tparquet::Encoding::RLE_DICTIONARY:
-        switch (type) {
-        case tparquet::Type::BOOLEAN:
-            if (encoding != tparquet::Encoding::PLAIN) {
-                return Status::InternalError("Bool type can't has dictionary page");
-            }
-        case tparquet::Type::BYTE_ARRAY:
-            decoder.reset(new ByteArrayDictDecoder());
-            break;
-        case tparquet::Type::INT32:
-            decoder.reset(new FixLengthDictDecoder<Int32>(type));
-            break;
-        case tparquet::Type::INT64:
-            decoder.reset(new FixLengthDictDecoder<Int64>(type));
-            break;
-        case tparquet::Type::INT96:
-            decoder.reset(new FixLengthDictDecoder<ParquetInt96>(type));
-            break;
-        case tparquet::Type::FLOAT:
-            decoder.reset(new FixLengthDictDecoder<Float32>(type));
-            break;
-        case tparquet::Type::DOUBLE:
-            decoder.reset(new FixLengthDictDecoder<Float64>(type));
-            break;
-        case tparquet::Type::FIXED_LEN_BYTE_ARRAY:
-            decoder.reset(new FixLengthDictDecoder<char*>(type));
-            break;
-        default:
-            return Status::InternalError("Unsupported type {}(encoding={}) in parquet decoder",
-                                         tparquet::to_string(type), tparquet::to_string(encoding));
-        }
-        break;
-    default:
-        return Status::InternalError("Unsupported encoding {}(type={}) in parquet decoder",
-                                     tparquet::to_string(encoding), tparquet::to_string(type));
-    }
-    return Status::OK();
-}
-
-void Decoder::init(FieldSchema* field_schema, cctz::time_zone* ctz) {
-    _field_schema = field_schema;
-    if (_decode_params == nullptr) {
-        _decode_params.reset(new DecodeParams());
-    }
-    if (ctz != nullptr) {
-        _decode_params->ctz = ctz;
-    }
-    const auto& schema = field_schema->parquet_schema;
-    if (schema.__isset.logicalType && schema.logicalType.__isset.TIMESTAMP) {
-        const auto& timestamp_info = schema.logicalType.TIMESTAMP;
-        if (!timestamp_info.isAdjustedToUTC) {
-            // should set timezone to utc+0
-            _decode_params->ctz = const_cast<cctz::time_zone*>(&_decode_params->utc0);
-        }
-        const auto& time_unit = timestamp_info.unit;
-        if (time_unit.__isset.MILLIS) {
-            _decode_params->second_mask = 1000;
-            _decode_params->scale_to_nano_factor = 1000000;
-        } else if (time_unit.__isset.MICROS) {
-            _decode_params->second_mask = 1000000;
-            _decode_params->scale_to_nano_factor = 1000;
-        } else if (time_unit.__isset.NANOS) {
-            _decode_params->second_mask = 1000000000;
-            _decode_params->scale_to_nano_factor = 1;
-        }
-    } else if (schema.__isset.converted_type) {
-        const auto& converted_type = schema.converted_type;
-        if (converted_type == tparquet::ConvertedType::TIMESTAMP_MILLIS) {
-            _decode_params->second_mask = 1000;
-            _decode_params->scale_to_nano_factor = 1000000;
-        } else if (converted_type == tparquet::ConvertedType::TIMESTAMP_MICROS) {
-            _decode_params->second_mask = 1000000;
-            _decode_params->scale_to_nano_factor = 1000;
-        }
-    }
-}
-
-void FixLengthDecoder::set_data(Slice* data) {
-    _data = data;
-    _offset = 0;
-}
-
-Status FixLengthDecoder::skip_values(size_t num_values) {
-    _offset += _type_length * num_values;
-    if (UNLIKELY(_offset > _data->size)) {
-        return Status::IOError("Out-of-bounds access in parquet data decoder");
-    }
-    return Status::OK();
-}
-
-Status FixLengthDecoder::decode_values(MutableColumnPtr& doris_column, DataTypePtr& data_type,
-                                       ColumnSelectVector& select_vector) {
-    size_t non_null_size = select_vector.num_values() - select_vector.num_nulls();
-    if (UNLIKELY(_offset + _type_length * non_null_size > _data->size)) {
-        return Status::IOError("Out-of-bounds access in parquet data decoder");
-    }
-    TypeIndex logical_type = remove_nullable(data_type)->get_type_id();
-    switch (logical_type) {
-#define DISPATCH(NUMERIC_TYPE, CPP_NUMERIC_TYPE, PHYSICAL_TYPE) \
-    case NUMERIC_TYPE:                                          \
-        return _decode_numeric<CPP_NUMERIC_TYPE>(doris_column, select_vector);
-        FOR_LOGICAL_NUMERIC_TYPES(DISPATCH)
-#undef DISPATCH
-    case TypeIndex::Date:
-        if (_physical_type == tparquet::Type::INT32) {
-            return _decode_date<VecDateTimeValue, Int64>(doris_column, select_vector);
-        }
-        break;
-    case TypeIndex::DateV2:
-        if (_physical_type == tparquet::Type::INT32) {
-            return _decode_date<DateV2Value<DateV2ValueType>, UInt32>(doris_column, select_vector);
-        }
-        break;
-    case TypeIndex::DateTime:
-        if (_physical_type == tparquet::Type::INT96) {
-            return _decode_datetime96<VecDateTimeValue, Int64>(doris_column, select_vector);
-        } else if (_physical_type == tparquet::Type::INT64) {
-            return _decode_datetime64<VecDateTimeValue, Int64>(doris_column, select_vector);
-        }
-        break;
-    case TypeIndex::DateTimeV2:
-        // Spark can set the timestamp precision by the following configuration:
-        // spark.sql.parquet.outputTimestampType = INT96(NANOS), TIMESTAMP_MICROS, TIMESTAMP_MILLIS
-        if (_physical_type == tparquet::Type::INT96) {
-            return _decode_datetime96<DateV2Value<DateTimeV2ValueType>, UInt64>(doris_column,
-                                                                                select_vector);
-        } else if (_physical_type == tparquet::Type::INT64) {
-            return _decode_datetime64<DateV2Value<DateTimeV2ValueType>, UInt64>(doris_column,
-                                                                                select_vector);
-        }
-        break;
-    case TypeIndex::Decimal32:
-        if (_physical_type == tparquet::Type::FIXED_LEN_BYTE_ARRAY) {
-            return _decode_binary_decimal<Int32>(doris_column, data_type, select_vector);
-        } else if (_physical_type == tparquet::Type::INT32) {
-            return _decode_primitive_decimal<Int32, Int32>(doris_column, data_type, select_vector);
-        } else if (_physical_type == tparquet::Type::INT64) {
-            return _decode_primitive_decimal<Int32, Int64>(doris_column, data_type, select_vector);
-        }
-        break;
-    case TypeIndex::Decimal64:
-        if (_physical_type == tparquet::Type::FIXED_LEN_BYTE_ARRAY) {
-            return _decode_binary_decimal<Int64>(doris_column, data_type, select_vector);
-        } else if (_physical_type == tparquet::Type::INT32) {
-            return _decode_primitive_decimal<Int64, Int32>(doris_column, data_type, select_vector);
-        } else if (_physical_type == tparquet::Type::INT64) {
-            return _decode_primitive_decimal<Int64, Int64>(doris_column, data_type, select_vector);
-        }
-        break;
-    case TypeIndex::Decimal128:
-        if (_physical_type == tparquet::Type::FIXED_LEN_BYTE_ARRAY) {
-            return _decode_binary_decimal<Int128>(doris_column, data_type, select_vector);
-        } else if (_physical_type == tparquet::Type::INT32) {
-            return _decode_primitive_decimal<Int128, Int32>(doris_column, data_type, select_vector);
-        } else if (_physical_type == tparquet::Type::INT64) {
-            return _decode_primitive_decimal<Int128, Int64>(doris_column, data_type, select_vector);
-        }
-        break;
-    case TypeIndex::Decimal128I:
-        if (_physical_type == tparquet::Type::FIXED_LEN_BYTE_ARRAY) {
-            return _decode_binary_decimal<Int128>(doris_column, data_type, select_vector);
-        } else if (_physical_type == tparquet::Type::INT32) {
-            return _decode_primitive_decimal<Int128, Int32>(doris_column, data_type, select_vector);
-        } else if (_physical_type == tparquet::Type::INT64) {
-            return _decode_primitive_decimal<Int128, Int64>(doris_column, data_type, select_vector);
-        }
-        break;
-    case TypeIndex::String:
-    case TypeIndex::FixedString:
-        if (_physical_type == tparquet::Type::FIXED_LEN_BYTE_ARRAY) {
-            return _decode_string(doris_column, select_vector);
-        }
-        break;
-    default:
-        break;
-    }
-
-    return Status::InvalidArgument("Can't decode parquet physical type {} to doris logical type {}",
-                                   tparquet::to_string(_physical_type), getTypeName(logical_type));
-}
-
-Status FixLengthDecoder::_decode_string(MutableColumnPtr& doris_column,
-                                        ColumnSelectVector& select_vector) {
-    ColumnSelectVector::DataReadType read_type;
-    while (size_t run_length = select_vector.get_next_run(&read_type)) {
-        switch (read_type) {
-        case ColumnSelectVector::CONTENT: {
-            std::vector<StringRef> string_values;
-            string_values.reserve(run_length);
-            for (size_t i = 0; i < run_length; ++i) {
-                char* buf_start = _data->data + _offset;
-                string_values.emplace_back(buf_start, _type_length);
-                _offset += _type_length;
-            }
-            doris_column->insert_many_strings(&string_values[0], run_length);
-            break;
-        }
-        case ColumnSelectVector::NULL_DATA: {
-            doris_column->insert_many_defaults(run_length);
-            break;
-        }
-        case ColumnSelectVector::FILTERED_CONTENT: {
-            _offset += _type_length * run_length;
-            break;
-        }
-        case ColumnSelectVector::FILTERED_NULL: {
-            // do nothing
-            break;
-        }
-        }
-    }
-    return Status::OK();
-}
-
-void ByteArrayDecoder::set_data(Slice* data) {
-    _data = data;
-    _offset = 0;
-}
-
-Status ByteArrayDecoder::skip_values(size_t num_values) {
-    for (int i = 0; i < num_values; ++i) {
-        if (UNLIKELY(_offset + 4 > _data->size)) {
-            return Status::IOError("Can't read byte array length from plain decoder");
-        }
-        uint32_t length =
-                decode_fixed32_le(reinterpret_cast<const uint8_t*>(_data->data) + _offset);
-        _offset += 4;
-        if (UNLIKELY(_offset + length) > _data->size) {
-            return Status::IOError("Can't skip enough bytes in plain decoder");
-        }
-        _offset += length;
-    }
-    return Status::OK();
-}
-
-Status ByteArrayDecoder::decode_values(MutableColumnPtr& doris_column, DataTypePtr& data_type,
-                                       ColumnSelectVector& select_vector) {
-    TypeIndex logical_type = remove_nullable(data_type)->get_type_id();
-    switch (logical_type) {
-    case TypeIndex::String:
-    case TypeIndex::FixedString: {
-        ColumnSelectVector::DataReadType read_type;
-        while (size_t run_length = select_vector.get_next_run(&read_type)) {
-            switch (read_type) {
-            case ColumnSelectVector::CONTENT: {
-                std::vector<StringRef> string_values;
-                string_values.reserve(run_length);
-                for (size_t i = 0; i < run_length; ++i) {
-                    if (UNLIKELY(_offset + 4 > _data->size)) {
-                        return Status::IOError("Can't read byte array length from plain decoder");
-                    }
-                    uint32_t length = decode_fixed32_le(
-                            reinterpret_cast<const uint8_t*>(_data->data) + _offset);
-                    _offset += 4;
-                    if (UNLIKELY(_offset + length) > _data->size) {
-                        return Status::IOError("Can't read enough bytes in plain decoder");
-                    }
-                    string_values.emplace_back(_data->data + _offset, length);
-                    _offset += length;
-                }
-                doris_column->insert_many_strings(&string_values[0], run_length);
-                break;
-            }
-            case ColumnSelectVector::NULL_DATA: {
-                doris_column->insert_many_defaults(run_length);
-                break;
-            }
-            case ColumnSelectVector::FILTERED_CONTENT: {
-                for (int i = 0; i < run_length; ++i) {
-                    if (UNLIKELY(_offset + 4 > _data->size)) {
-                        return Status::IOError("Can't read byte array length from plain decoder");
-                    }
-                    uint32_t length = decode_fixed32_le(
-                            reinterpret_cast<const uint8_t*>(_data->data) + _offset);
-                    _offset += 4;
-                    if (UNLIKELY(_offset + length) > _data->size) {
-                        return Status::IOError("Can't read enough bytes in plain decoder");
-                    }
-                    _offset += length;
-                }
-                break;
-            }
-            case ColumnSelectVector::FILTERED_NULL: {
-                // do nothing
-                break;
-            }
-            }
-        }
-        return Status::OK();
-    }
-    case TypeIndex::Decimal32:
-        return _decode_binary_decimal<Int32>(doris_column, data_type, select_vector);
-    case TypeIndex::Decimal64:
-        return _decode_binary_decimal<Int64>(doris_column, data_type, select_vector);
-    case TypeIndex::Decimal128:
-        return _decode_binary_decimal<Int128>(doris_column, data_type, select_vector);
-    case TypeIndex::Decimal128I:
-        return _decode_binary_decimal<Int128>(doris_column, data_type, select_vector);
-    default:
-        break;
-    }
-    return Status::InvalidArgument(
-            "Can't decode parquet physical type BYTE_ARRAY to doris logical type {}",
-            getTypeName(logical_type));
-}
-
-void ByteArrayDictDecoder::set_data(Slice* data) {
-    _data = data;
-    _offset = 0;
-    uint8_t bit_width = *data->data;
-    _index_batch_decoder.reset(
-            new RleBatchDecoder<uint32_t>(reinterpret_cast<uint8_t*>(data->data) + 1,
-                                          static_cast<int>(data->size) - 1, bit_width));
-}
-
-Status ByteArrayDictDecoder::set_dict(std::unique_ptr<uint8_t[]>& dict, int32_t length,
-                                      size_t num_values) {
-    _dict = std::move(dict);
-    _dict_items.reserve(num_values);
-    uint32_t offset_cursor = 0;
-    char* dict_item_address = reinterpret_cast<char*>(_dict.get());
-
-    size_t total_length = 0;
-    for (int i = 0; i < num_values; ++i) {
-        uint32_t l = decode_fixed32_le(_dict.get() + offset_cursor);
-        offset_cursor += 4;
-        offset_cursor += l;
-        total_length += l;
-    }
-
-    // For insert_many_strings_overflow
-    _dict_data.resize(total_length + MAX_STRINGS_OVERFLOW_SIZE);
-    _max_value_length = 0;
-    size_t offset = 0;
-    offset_cursor = 0;
-    for (int i = 0; i < num_values; ++i) {
-        uint32_t l = decode_fixed32_le(_dict.get() + offset_cursor);
-        offset_cursor += 4;
-        memcpy(&_dict_data[offset], dict_item_address + offset_cursor, l);
-        _dict_items.emplace_back(&_dict_data[offset], l);
-        offset_cursor += l;
-        offset += l;
-        if (offset_cursor > length) {
-            return Status::Corruption("Wrong data length in dictionary");
-        }
-        if (l > _max_value_length) {
-            _max_value_length = l;
-        }
-    }
-    if (offset_cursor != length) {
-        return Status::Corruption("Wrong dictionary data for byte array type");
-    }
-    return Status::OK();
-}
-
-Status ByteArrayDictDecoder::skip_values(size_t num_values) {
-    _indexes.resize(num_values);
-    _index_batch_decoder->GetBatch(&_indexes[0], num_values);
-    return Status::OK();
-}
-
-Status ByteArrayDictDecoder::decode_values(MutableColumnPtr& doris_column, DataTypePtr& data_type,
-                                           ColumnSelectVector& select_vector) {
-    size_t non_null_size = select_vector.num_values() - select_vector.num_nulls();
-    if (doris_column->is_column_dictionary() &&
-        assert_cast<ColumnDictI32&>(*doris_column).dict_size() == 0) {
-        assert_cast<ColumnDictI32&>(*doris_column)
-                .insert_many_dict_data(&_dict_items[0], _dict_items.size());
-    }
-    _indexes.resize(non_null_size);
-    _index_batch_decoder->GetBatch(&_indexes[0], non_null_size);
-
-    if (doris_column->is_column_dictionary()) {
-        return _decode_dict_values(doris_column, select_vector);
-    }
-
-    TypeIndex logical_type = remove_nullable(data_type)->get_type_id();
-    switch (logical_type) {
-    case TypeIndex::String:
-    case TypeIndex::FixedString: {
-        size_t dict_index = 0;
-
-        ColumnSelectVector::DataReadType read_type;
-        while (size_t run_length = select_vector.get_next_run(&read_type)) {
-            switch (read_type) {
-            case ColumnSelectVector::CONTENT: {
-                std::vector<StringRef> string_values;
-                string_values.reserve(run_length);
-                for (size_t i = 0; i < run_length; ++i) {
-                    string_values.emplace_back(_dict_items[_indexes[dict_index++]]);
-                }
-                doris_column->insert_many_strings_overflow(&string_values[0], run_length,
-                                                           _max_value_length);
-                break;
-            }
-            case ColumnSelectVector::NULL_DATA: {
-                doris_column->insert_many_defaults(run_length);
-                break;
-            }
-            case ColumnSelectVector::FILTERED_CONTENT: {
-                dict_index += run_length;
-                break;
-            }
-            case ColumnSelectVector::FILTERED_NULL: {
-                // do nothing
-                break;
-            }
-            }
-        }
-        return Status::OK();
-    }
-    case TypeIndex::Decimal32:
-        return _decode_binary_decimal<Int32>(doris_column, data_type, select_vector);
-    case TypeIndex::Decimal64:
-        return _decode_binary_decimal<Int64>(doris_column, data_type, select_vector);
-    case TypeIndex::Decimal128:
-        return _decode_binary_decimal<Int128>(doris_column, data_type, select_vector);
-    case TypeIndex::Decimal128I:
-        return _decode_binary_decimal<Int128>(doris_column, data_type, select_vector);
-    default:
-        break;
-    }
-    return Status::InvalidArgument(
-            "Can't decode parquet physical type BYTE_ARRAY to doris logical type {}",
-            getTypeName(logical_type));
-}
-
-Status ByteArrayDictDecoder::_decode_dict_values(MutableColumnPtr& doris_column,
-                                                 ColumnSelectVector& select_vector) {
-    DCHECK(doris_column->is_column_dictionary());
-    size_t dict_index = 0;
-    ColumnSelectVector::DataReadType read_type;
-    auto& column_data = assert_cast<ColumnDictI32&>(*doris_column).get_data();
-    while (size_t run_length = select_vector.get_next_run(&read_type)) {
-        switch (read_type) {
-        case ColumnSelectVector::CONTENT: {
-            uint32_t* start_index = &_indexes[0];
-            column_data.insert(start_index + dict_index, start_index + dict_index + run_length);
-            dict_index += run_length;
-            break;
-        }
-        case ColumnSelectVector::NULL_DATA: {
-            doris_column->insert_many_defaults(run_length);
-            break;
-        }
-        case ColumnSelectVector::FILTERED_CONTENT: {
-            dict_index += run_length;
-            break;
-        }
-        case ColumnSelectVector::FILTERED_NULL: {
-            break;
-        }
-        }
-    }
-    return Status::OK();
-}
-
-Status BoolPlainDecoder::skip_values(size_t num_values) {
-    int skip_cached = std::min(num_unpacked_values_ - unpacked_value_idx_, (int)num_values);
-    unpacked_value_idx_ += skip_cached;
-    if (skip_cached == num_values) {
-        return Status::OK();
-    }
-    int num_remaining = num_values - skip_cached;
-    int num_to_skip = BitUtil::RoundDownToPowerOf2(num_remaining, 32);
-    if (num_to_skip > 0) {
-        bool_values_.SkipBatch(1, num_to_skip);
-    }
-    num_remaining -= num_to_skip;
-    if (num_remaining > 0) {
-        DCHECK_LE(num_remaining, UNPACKED_BUFFER_LEN);
-        num_unpacked_values_ =
-                bool_values_.UnpackBatch(1, UNPACKED_BUFFER_LEN, &unpacked_values_[0]);
-        if (UNLIKELY(num_unpacked_values_ < num_remaining)) {
-            return Status::IOError("Can't skip enough booleans in plain decoder");
-        }
-        unpacked_value_idx_ = num_remaining;
-    }
-    return Status::OK();
-}
-
-Status BoolPlainDecoder::decode_values(MutableColumnPtr& doris_column, DataTypePtr& data_type,
-                                       ColumnSelectVector& select_vector) {
-    auto& column_data = static_cast<ColumnVector<UInt8>&>(*doris_column).get_data();
-    size_t data_index = column_data.size();
-    column_data.resize(data_index + select_vector.num_values() - select_vector.num_filtered());
-
-    ColumnSelectVector::DataReadType read_type;
-    while (size_t run_length = select_vector.get_next_run(&read_type)) {
-        switch (read_type) {
-        case ColumnSelectVector::CONTENT: {
-            bool value;
-            for (size_t i = 0; i < run_length; ++i) {
-                if (UNLIKELY(!_decode_value(&value))) {
-                    return Status::IOError("Can't read enough booleans in plain decoder");
-                }
-                column_data[data_index++] = (UInt8)value;
-            }
-            break;
-        }
-        case ColumnSelectVector::NULL_DATA: {
-            data_index += run_length;
-            break;
-        }
-        case ColumnSelectVector::FILTERED_CONTENT: {
-            bool value;
-            for (int i = 0; i < run_length; ++i) {
-                if (UNLIKELY(!_decode_value(&value))) {
-                    return Status::IOError("Can't read enough booleans in plain decoder");
-                }
-            }
-            break;
-        }
-        case ColumnSelectVector::FILTERED_NULL: {
-            // do nothing
-            break;
-        }
-        }
-    }
-    return Status::OK();
 }
 } // namespace doris::vectorized
