+++ be/src/vec/exec/format/parquet/parquet_pred_cmp.h
@@ -21,7 +21,10 @@
 #include <vector>
 
 #include "exec/olap_common.h"
+#include "gutil/endian.h"
 #include "parquet_common.h"
+#include "vec/data_types/data_type_decimal.h"
+#include "vec/exec/format/format_common.h"
 
 namespace doris::vectorized {
