+++ be/src/vec/exec/format/table/iceberg_reader.h
@@ -21,6 +21,7 @@
 
 #include "table_format_reader.h"
 #include "vec/columns/column_dictionary.h"
+#include "vec/exec/format/format_common.h"
 #include "vec/exec/format/generic_reader.h"
 #include "vec/exec/format/parquet/parquet_common.h"
 #include "vec/exprs/vexpr.h"
